create useOpenEmailThreadInCommandMenu
