+++ packages/twenty-front/src/modules/command-menu/hooks/useCommandMenu.ts
@@ -1,15 +1,13 @@
 import { useRecoilCallback } from 'recoil';
 
 import { commandMenuSearchState } from '@/command-menu/states/commandMenuSearchState';
-import { IconDotsVertical, IconMail } from 'twenty-ui';
+import { IconDotsVertical } from 'twenty-ui';
 
 import { useNavigateCommandMenu } from '@/command-menu/hooks/useNavigateCommandMenu';
-import { viewableRecordIdComponentState } from '@/command-menu/pages/record-page/states/viewableRecordIdComponentState';
 import { isCommandMenuClosingState } from '@/command-menu/states/isCommandMenuClosingState';
 import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
 import { isDragSelectionStartEnabledState } from '@/ui/utilities/drag-select/states/internal/isDragSelectionStartEnabledState';
 import { useCallback } from 'react';
-import { v4 } from 'uuid';
 import { isCommandMenuOpenedState } from '../states/isCommandMenuOpenedState';
 
 export const useCommandMenu = () => {
@@ -53,60 +51,11 @@ export const useCommandMenu = () => {
     [closeCommandMenu, openCommandMenu],
   );
 
-  const openEmailThreadInCommandMenu = useRecoilCallback(
-    ({ set }) => {
-      return (emailThreadId: string) => {
-        const pageComponentInstanceId = v4();
-
-        set(
-          viewableRecordIdComponentState.atomFamily({
-            instanceId: pageComponentInstanceId,
-          }),
-          emailThreadId,
-        );
-
-        // TODO: Uncomment this once we need to show the thread title in the navigation
-        // const objectMetadataItem = snapshot
-        //   .getLoadable(objectMetadataItemsState)
-        //   .getValue()
-        //   .find(
-        //     ({ nameSingular }) =>
-        //       nameSingular === CoreObjectNameSingular.MessageThread,
-        //   );
-
-        // set(
-        //   commandMenuNavigationMorphItemsState,
-        //   new Map([
-        //     ...snapshot
-        //       .getLoadable(commandMenuNavigationMorphItemsState)
-        //       .getValue(),
-        //     [
-        //       pageComponentInstanceId,
-        //       {
-        //         objectMetadataId: objectMetadataItem?.id,
-        //         recordId: emailThreadId,
-        //       },
-        //     ],
-        //   ]),
-        // );
-
-        navigateCommandMenu({
-          page: CommandMenuPages.ViewEmailThread,
-          pageTitle: 'Email Thread',
-          pageIcon: IconMail,
-          pageId: pageComponentInstanceId,
-        });
-      };
-    },
-    [navigateCommandMenu],
-  );
-
   return {
     openCommandMenu,
     closeCommandMenu,
     onCommandMenuCloseAnimationComplete,
     navigateCommandMenu,
     toggleCommandMenu,
-    openEmailThreadInCommandMenu,
   };
 };
