+++ packages/twenty-front/src/modules/command-menu/hooks/useOpenEmailThreadInCommandMenu.ts
@@ -1,0 +1,62 @@
+import { useNavigateCommandMenu } from '@/command-menu/hooks/useNavigateCommandMenu';
+import { viewableRecordIdComponentState } from '@/command-menu/pages/record-page/states/viewableRecordIdComponentState';
+import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
+import { useRecoilCallback } from 'recoil';
+import { IconMail } from 'twenty-ui';
+import { v4 } from 'uuid';
+
+export const useOpenEmailThreadInCommandMenu = () => {
+  const { navigateCommandMenu } = useNavigateCommandMenu();
+
+  const openEmailThreadInCommandMenu = useRecoilCallback(
+    ({ set }) => {
+      return (emailThreadId: string) => {
+        const pageComponentInstanceId = v4();
+
+        set(
+          viewableRecordIdComponentState.atomFamily({
+            instanceId: pageComponentInstanceId,
+          }),
+          emailThreadId,
+        );
+
+        // TODO: Uncomment this once we need to show the thread title in the navigation
+        // const objectMetadataItem = snapshot
+        //   .getLoadable(objectMetadataItemsState)
+        //   .getValue()
+        //   .find(
+        //     ({ nameSingular }) =>
+        //       nameSingular === CoreObjectNameSingular.MessageThread,
+        //   );
+
+        // set(
+        //   commandMenuNavigationMorphItemsState,
+        //   new Map([
+        //     ...snapshot
+        //       .getLoadable(commandMenuNavigationMorphItemsState)
+        //       .getValue(),
+        //     [
+        //       pageComponentInstanceId,
+        //       {
+        //         objectMetadataId: objectMetadataItem?.id,
+        //         recordId: emailThreadId,
+        //       },
+        //     ],
+        //   ]),
+        // );
+
+        navigateCommandMenu({
+          page: CommandMenuPages.ViewEmailThread,
+          pageTitle: 'Email Thread',
+          pageIcon: IconMail,
+          pageId: pageComponentInstanceId,
+        });
+      };
+    },
+    [navigateCommandMenu],
+  );
+
+  return {
+    openEmailThreadInCommandMenu,
+  };
+};
